requirements.txt
django==6.0.4
djangorestframework==3.17.1
django-filter==25.2
markdown==3.10.2
psycopg[binary]==3.2.10
gunicorn==23.0.0

